refactor: Main is not separate class any more, classes moved to separate packages
