--- src/pl/edu/zut/wdsi/Main.java
@@ -1,10 +1,0 @@
-/* This is develop branch */
-package pl.edu.zut.wdsi;
-
-public class Main {
-    public static void main(String[] args) {
-
-        Puzzle p = new Puzzle();
-        // write your code here
-    }
-}
