+++ src/pl/edu/zut/wdsi/puzzle/Moves.java
@@ -1,4 +1,4 @@
-package pl.edu.zut.wdsi;
+package pl.edu.zut.wdsi.puzzle;
 
 public enum Moves {
     GOUP,
